requirements.txt
flask==2.3.3
tensorflow==2.15.0
numpy==1.24.3
pillow==10.0.0
gunicorn==21.2.0
werkzeug==2.3.7
